requirements.txt
basicsr
facexlib
transformers
addict
future
lmdb
numpy
opencv-python-headless
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch>=1.7.1
torchvision
tqdm
yapf
lpips
boto3
sanic
diffusers
